Split BigSampler for formats into separate files
