+++ ratatool-sampling/src/main/scala/com/spotify/ratatool/samplers/BigSampler.scala
@@ -16,38 +16,26 @@
  */
 package com.spotify.ratatool.samplers
 
-import java.io.IOException
 import java.net.URI
 import java.nio.charset.Charset
-import java.util.{List => JList}
 
-import com.google.api.services.bigquery.model.{TableFieldSchema, TableReference, TableSchema}
+import com.google.api.services.bigquery.model.{TableFieldSchema, TableReference}
 import com.google.common.hash.{HashCode, Hasher, Hashing}
-import com.google.common.io.BaseEncoding
-import com.google.protobuf.{AbstractMessage, ByteString}
-import com.google.protobuf.Descriptors.FieldDescriptor.JavaType
 import com.spotify.ratatool.samplers.util.SamplerSCollectionFunctions._
 import com.spotify.ratatool.Command
 import com.spotify.ratatool.avro.specific.TestRecord
-import com.spotify.ratatool.io.{AvroIO, FileStorage}
 import com.spotify.ratatool.samplers.util._
-import com.spotify.ratatool.serde.JsonSerDe
 import com.spotify.scio.bigquery.TableRow
-import com.spotify.scio.io.{Tap, Taps}
+import com.spotify.scio.io.Tap
 import com.spotify.scio.values.SCollection
 import com.spotify.scio.{ContextAndArgs, ScioContext}
 import org.apache.avro.Schema
-import org.apache.avro.Schema.Type
-import org.apache.avro.generic.GenericRecord
 import org.apache.beam.runners.dataflow.options.DataflowPipelineWorkerPoolOptions
 import org.apache.beam.sdk.io.FileSystems
-import org.apache.beam.sdk.io.gcp.bigquery.{BigQueryHelpers, BigQueryIO, BigQueryOptions,
-  PatchedBigQueryServicesImpl}
+import org.apache.beam.sdk.io.gcp.bigquery.BigQueryHelpers
 import org.apache.beam.sdk.options.PipelineOptions
 import org.slf4j.LoggerFactory
 
-import scala.annotation.tailrec
-import scala.collection.JavaConverters._
 import scala.concurrent.Future
 import scala.language.existentials
 import scala.util.Try
@@ -346,444 +334,4 @@ object BigSampler extends Command {
   }
 }
 
-private[samplers] object BigSamplerProtobuf {
-  private val log = LoggerFactory.getLogger(BigSamplerProtobuf.getClass)
 
-  private[samplers] def buildKey(distributionFields: Seq[String])(m: AbstractMessage)
-  : Set[String] = {
-    distributionFields.map(f => getProtobufField(m, f).toString).toSet
-  }
-
-  // scalastyle:off cyclomatic.complexity
-  private[samplers] def hashProtobufField[T <: AbstractMessage : ClassTag](m: T,
-                                                                           fieldStr: String,
-                                                                           hasher: Hasher)
-  : Hasher = {
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = Option(m.getDescriptorForType.findFieldByName(subfields.head)).getOrElse {
-      throw new NoSuchElementException(s"Can't find field $fieldStr in protobuf schema")
-    }
-    val v = m.getField(field)
-    if (v == null) {
-      log.debug(s"Field `${field.getFullName}` of type ${field.getType} is null - won't account" +
-      s" for hash")
-      hasher
-    } else {
-      field.getJavaType match {
-        case JavaType.MESSAGE => hashProtobufField(
-          v.asInstanceOf[AbstractMessage], subfields.tail.mkString("."), hasher)
-        case JavaType.INT => hasher.putLong(v.asInstanceOf[Int].toLong)
-        case JavaType.LONG => hasher.putLong(v.asInstanceOf[Long])
-        case JavaType.FLOAT => hasher.putFloat(v.asInstanceOf[Float])
-        case JavaType.DOUBLE => hasher.putDouble(v.asInstanceOf[Double])
-        case JavaType.BOOLEAN => hasher.putBoolean(v.asInstanceOf[Boolean])
-        case JavaType.STRING => hasher.putString(v.asInstanceOf[CharSequence],
-          BigSampler.utf8Charset)
-        case JavaType.BYTE_STRING => hasher.putBytes(v.asInstanceOf[ByteString].toByteArray)
-        case JavaType.ENUM => hasher.putString(v.asInstanceOf[Enum[_]].name, BigSampler.utf8Charset)
-        // Array, Union
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  // scalastyle:off cyclomatic.complexity
-  private[samplers] def getProtobufField[T <: AbstractMessage : ClassTag](m: T, fieldStr: String)
-  : Any = {
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = Option(m.getDescriptorForType.findFieldByName(subfields.head)).getOrElse {
-      throw new NoSuchElementException(s"Can't find field $fieldStr in protobuf schema")
-    }
-    val v = m.getField(field)
-    if (v == null) {
-      log.debug(s"Field `${field.getFullName}` of type ${field.getType} is null - won't account" +
-        s" for key")
-    } else {
-      field.getJavaType match {
-        case JavaType.MESSAGE => getProtobufField(
-          v.asInstanceOf[AbstractMessage], subfields.tail.mkString("."))
-        case JavaType.INT => v.asInstanceOf[Int]
-        case JavaType.LONG => v.asInstanceOf[Long]
-        case JavaType.FLOAT => v.asInstanceOf[Float]
-        case JavaType.DOUBLE => v.asInstanceOf[Double]
-        case JavaType.BOOLEAN => v.asInstanceOf[Boolean]
-        case JavaType.STRING => v.asInstanceOf[CharSequence]
-        case JavaType.BYTE_STRING => v.asInstanceOf[ByteString].toByteArray
-        case JavaType.ENUM => v.asInstanceOf[Enum[_]].name
-        // Array, Union
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-
-}
-
-private[samplers] object BigSamplerAvro {
-  private val log = LoggerFactory.getLogger(BigSamplerAvro.getClass)
-
-  // scalastyle:off cyclomatic.complexity
-  @tailrec
-  private[samplers] def hashAvroField(schema: Schema)(r: GenericRecord,
-                                      fieldStr: String,
-                                      hasher: Hasher): Hasher = {
-    val recordSchema = getRecordSchema(schema)
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = recordSchema.getFields.asScala.find(_.name == subfields.head).getOrElse {
-      throw new NoSuchElementException(s"Can't find field $fieldStr in avro schema $schema")
-    }
-    val v = r.get(subfields.head)
-    if (v == null) {
-      log.debug(s"Field `${field.name}` of type ${field.schema.getType} is null - won't account" +
-        s" for hash")
-      hasher
-    } else {
-      field.schema.getType match {
-        case Type.RECORD =>
-          hashAvroField(field.schema)(
-            v.asInstanceOf[GenericRecord],
-            subfields.tail.mkString(BigSampler.fieldSep.toString),
-            hasher)
-        case Type.ENUM => hasher.putString(v.asInstanceOf[Enum[_]].name, BigSampler.utf8Charset)
-        case Type.STRING => hasher.putString(v.asInstanceOf[CharSequence], BigSampler.utf8Charset)
-        case Type.BYTES => hasher.putBytes(v.asInstanceOf[Array[Byte]])
-        // to keep it consistent with BigQuery INT - convert int to long
-        case Type.INT => hasher.putLong(v.asInstanceOf[Int].toLong)
-        case Type.LONG => hasher.putLong(v.asInstanceOf[Long])
-        case Type.FLOAT => hasher.putFloat(v.asInstanceOf[Float])
-        case Type.DOUBLE => hasher.putDouble(v.asInstanceOf[Double])
-        case Type.BOOLEAN => hasher.putBoolean(v.asInstanceOf[Boolean])
-        case Type.UNION => hashAvroUnionField(field, v, hasher)
-        case Type.ARRAY => hashAvroArrayField(field, v, hasher)
-        //  Type.MAP | Type.FIXED =>
-        case t => throw new UnsupportedOperationException(
-          s"Type `${field.schema.getType}` of `${field.name}` is not supported as sampling key!")
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  // scalastyle:off cyclomatic.complexity
-  private def hashAvroArrayField(field: Schema.Field, v: AnyRef, hasher: Hasher): Hasher = {
-    field.schema.getElementType.getType match {
-      case Type.ENUM => v.asInstanceOf[JList[Enum[_]]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putString(e.name, BigSampler.utf8Charset))
-      case Type.STRING => v.asInstanceOf[JList[CharSequence]].asScala.foldLeft(hasher)(
-        (hasher, e) => hasher.putString(e, BigSampler.utf8Charset))
-      case Type.BYTES => v.asInstanceOf[JList[Array[Byte]]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putBytes(e))
-      case Type.INT => v.asInstanceOf[JList[Int]].asScala.foldLeft(hasher)((hasher, e) =>
-        // to keep it consistent with BigQuery INT - convert int to long
-        hasher.putLong(e.toLong))
-      case Type.LONG => v.asInstanceOf[JList[Long]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putLong(e))
-      case Type.FLOAT => v.asInstanceOf[JList[Float]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putFloat(e))
-      case Type.DOUBLE => v.asInstanceOf[JList[Double]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putDouble(e))
-      case Type.BOOLEAN => v.asInstanceOf[JList[Boolean]].asScala.foldLeft(hasher)((hasher, e) =>
-        hasher.putBoolean(e))
-      case Type.NULL =>
-        // ignore nulls
-        hasher
-      case _ => throw new UnsupportedOperationException(
-        s"Type ${field.schema.getElementType.getType} is not supported as hash for array ")
-    }
-  }
-
-  // scalastyle:off cyclomatic.complexity
-  private def hashAvroUnionField(field: Schema.Field, v: AnyRef, hasher: Hasher): Hasher = {
-    val types = field.schema.getTypes.asScala
-    types.foldLeft(hasher)( (hasher, p) =>
-      p.getType match {
-        case Type.ENUM if v.isInstanceOf[Enum[_]] =>
-          hasher.putString(v.asInstanceOf[Enum[_]].name, BigSampler.utf8Charset)
-        case Type.STRING if v.isInstanceOf[CharSequence] =>
-          hasher.putString(v.asInstanceOf[CharSequence], BigSampler.utf8Charset)
-        case Type.BYTES if v.isInstanceOf[Array[Byte]] =>
-          hasher.putBytes(v.asInstanceOf[Array[Byte]])
-        // to keep it consistent with BigQuery INT - convert int to long
-        case Type.INT if v.isInstanceOf[Int] => hasher.putLong(v.asInstanceOf[Int].toLong)
-        case Type.LONG if v.isInstanceOf[Long] => hasher.putLong(v.asInstanceOf[Long])
-        case Type.FLOAT if v.isInstanceOf[Float] => hasher.putFloat(v.asInstanceOf[Float])
-        case Type.DOUBLE if v.isInstanceOf[Double] => hasher.putDouble(v.asInstanceOf[Double])
-        case Type.BOOLEAN if v.isInstanceOf[Boolean] =>
-          hasher.putBoolean(v.asInstanceOf[Boolean])
-        case Type.NULL =>
-          // ignore nulls
-          hasher
-        case _ => throw new UnsupportedOperationException(
-          s"Value `$v` of union ${field.name} has unsupported type `${p.getType}`")
-      }
-    )
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  private def getRecordSchema(schema: Schema): Schema = {
-    schema.getType match {
-      case Type.UNION => schema.getTypes.asScala.head
-      case Type.RECORD => schema
-      case _ => throw new IOException(s"Can't recognise schema `$schema` as record")
-    }
-  }
-
-  @tailrec
-  private def fieldInAvroSchema(schema: Schema, fieldStr: String): Boolean = {
-    val recordSchema = getRecordSchema(schema)
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val fieldOpt = recordSchema.getFields.asScala.find(_.name == subfields.head)
-    if (fieldOpt.isEmpty) {
-      false
-    } else {
-      val field = fieldOpt.get
-      field.schema.getType match {
-        case Type.RECORD =>
-          fieldInAvroSchema(field.schema, subfields.tail.mkString(BigSampler.fieldSep.toString))
-        case Type.ENUM | Type.STRING | Type.BYTES | Type.INT | Type.LONG| Type.FLOAT | Type.DOUBLE
-             | Type.BOOLEAN => true
-        case  t => throw new UnsupportedOperationException(
-          s"Type `${field.schema.getType}` of `${field.name}` is not supported as sampling key!")
-      }
-    }
-  }
-
-  // scalastyle:off cyclomatic.complexity
-  // TODO: Potentially reduce this and hashAvroField to a single function
-  @tailrec
-  private[samplers] def getAvroField(r: GenericRecord,
-                                      fieldStr: String,
-                                      schema: Schema): Any = {
-    val recordSchema = getRecordSchema(schema)
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = recordSchema.getFields.asScala.find(_.name == subfields.head).getOrElse {
-      throw new NoSuchElementException(s"Can't find field $fieldStr in avro schema $schema")
-    }
-    val v = r.get(subfields.head)
-    if (v == null) {
-      log.debug(s"Field `${field.name}` of type ${field.schema.getType} is null, will not look" +
-        " for nested values")
-      null
-    } else {
-      field.schema.getType match {
-        case Type.RECORD if fieldStr.nonEmpty =>
-          getAvroField(
-            v.asInstanceOf[GenericRecord],
-            subfields.tail.mkString(BigSampler.fieldSep.toString),
-            field.schema)
-        case Type.ENUM => v.asInstanceOf[Enum[_]]
-        case Type.STRING => v.asInstanceOf[CharSequence]
-        case Type.BYTES => v.asInstanceOf[Array[Byte]]
-        case Type.INT => v.asInstanceOf[Int].toLong
-        case Type.LONG => v.asInstanceOf[Long]
-        case Type.FLOAT => v.asInstanceOf[Float]
-        case Type.DOUBLE => v.asInstanceOf[Double]
-        case Type.BOOLEAN => v.asInstanceOf[Boolean]
-        //TODO: Support Union type
-        case _ => throw new Exception(s"Current type ${field.schema.getType} is unsupported " +
-          s"or field $fieldStr does not refer to a single field.")
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  private[samplers] def buildKey(schema: => Schema,
-                                 distributionFields: Seq[String])(gr: GenericRecord)
-  : Set[String] = {
-    distributionFields.map(f => getAvroField(gr, f, schema).toString).toSet
-  }
-
-  //scalastyle:off method.length cyclomatic.complexity parameter.number
-  private[samplers] def sample(sc: ScioContext,
-                               input: String,
-                               output: String,
-                               fields: Seq[String],
-                               fraction: Double,
-                               seed: Option[Int],
-                               distribution: Option[SampleDistribution],
-                               distributionFields: Seq[String],
-                               precision: Precision,
-                               maxKeySize: Int)
-  : Future[Tap[GenericRecord]] = {
-    val schema = AvroIO.getAvroSchemaFromFile(input)
-    val outputParts = if (output.endsWith("/")) output + "part*" else output + "/part*"
-    if (FileStorage(outputParts).isDone) {
-      log.info(s"Reuse previous sample at $outputParts")
-      Taps().avroFile(outputParts, schema = schema)
-    } else {
-      log.info(s"Will sample from: $input, output will be $output")
-      val schemaSer = schema.toString(false)
-      @transient lazy val schemaSerDe = new Schema.Parser().parse(schemaSer)
-
-      val coll = sc.avroFile[GenericRecord](input, schema)
-
-      val sampledCollection = sampleAvro(coll, fraction, schema, fields, seed, distribution,
-        distributionFields, precision, maxKeySize)
-
-
-      val r = sampledCollection.saveAsAvroFile(output, schema = schema)
-      sc.close().waitUntilDone()
-      r
-    }
-  }
-  //scalastyle:on method.length cyclomatic.complexity parameter.number
-}
-
-private[samplers] object BigSamplerBigQuery {
-
-  private val log = LoggerFactory.getLogger(BigSamplerBigQuery.getClass)
-
-  // scalastyle:off cyclomatic.complexity
-  private[samplers] def hashTableRow(tblSchema: => Seq[TableFieldSchema])(r: TableRow,
-                                     fieldStr: String,
-                                     hasher: Hasher): Hasher = {
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = tblSchema.find(_.getName == subfields.head).getOrElse {
-      throw new NoSuchElementException(s"Can't find field `$fieldStr` in the schema $tblSchema")
-    }
-    val v = r.get(subfields.head)
-    if (v == null) {
-      log.debug(s"Field `${field.getName}` of type ${field.getType} and mode ${field.getMode}" +
-        s" is null - won't account for hash")
-      hasher
-    } else {
-      val vs = if (field.getMode == "REPEATED") {
-        v.asInstanceOf[JList[AnyRef]].asScala
-      } else {
-        Seq(v)
-      }
-      field.getType match {
-        case "BOOLEAN" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putBoolean(v.toString.toBoolean))
-        case "INTEGER" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putLong(v.toString.toLong))
-        case "FLOAT" => vs.foldLeft(hasher)((hasher, v) => hasher.putFloat(v.toString.toFloat))
-        case "STRING" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putString(v.toString, BigSampler.utf8Charset))
-        case "BYTES" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putBytes(BaseEncoding.base64().decode(v.toString)))
-        case "TIMESTAMP" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putString(v.toString, BigSampler.utf8Charset))
-        case "DATE" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putString(v.toString, BigSampler.utf8Charset))
-        case "TIME" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putString(v.toString, BigSampler.utf8Charset))
-        case "DATETIME" => vs.foldLeft(hasher)((hasher, v) =>
-          hasher.putString(v.toString, BigSampler.utf8Charset))
-        case "RECORD" =>
-          vs.foldLeft(hasher)((hasher, vi) =>
-            hashTableRow(field.getFields.asScala)(
-              vi.asInstanceOf[TableRow],
-              subfields.tail.mkString(BigSampler.fieldSep.toString),
-              hasher)
-          )
-        case t => throw new UnsupportedOperationException(
-          s"Type `$t` of field `${field.getName}` is not supported as sampling key")
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  // scalastyle:off cyclomatic.complexity
-  // TODO: Potentially reduce this and hashAvroField to a single function
-  @tailrec
-  private[samplers] def getTableRowField(r: TableRow,
-                                         fieldStr: String,
-                                         tblSchema: Seq[TableFieldSchema]): Any = {
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val field = tblSchema.find(_.getName == subfields.head).getOrElse {
-      throw new NoSuchElementException(s"Can't find field `$fieldStr` in the schema $tblSchema")
-    }
-    val v = r.get(subfields.head)
-    if (v == null) {
-      log.debug(s"Field `${field.getName}` of type ${field.getType} and mode ${field.getMode}" +
-        s" is null - won't account for hash")
-    } else {
-      field.getType match {
-        case "BOOLEAN" => v.toString.toBoolean
-        case "INTEGER" => v.toString.toLong
-        case "FLOAT" => v.toString.toFloat
-        case "STRING" => v.toString
-        case "BYTES" => BaseEncoding.base64().decode(v.toString)
-        case "TIMESTAMP" => v.toString
-        case "DATE" => v.toString
-        case "TIME" => v.toString
-        case "DATETIME" => v.toString
-        case "RECORD" if fieldStr.nonEmpty =>
-          // v is a LinkedHashMap
-          getTableRowField(
-            TableRow(v.asInstanceOf[java.util.Map[String, AnyRef]].asScala.toList: _*),
-            subfields.tail.mkString(BigSampler.fieldSep.toString),
-            field.getFields.asScala)
-        case t => throw new UnsupportedOperationException(
-          s"Type `$t` of field `${field.getName}` is not supported as sampling key")
-      }
-    }
-  }
-  // scalastyle:on cyclomatic.complexity
-
-  @tailrec
-  private def fieldInTblSchema(tblSchema: Seq[TableFieldSchema], fieldStr: String): Boolean = {
-    val subfields = fieldStr.split(BigSampler.fieldSep)
-    val fieldOpt = tblSchema.find(_.getName == subfields.head)
-    if (fieldOpt.isEmpty) {
-      false
-    } else {
-      val field = fieldOpt.get
-      field.getType match {
-        case "RECORD" =>
-          fieldInTblSchema(
-            field.getFields.asScala,
-            subfields.tail.mkString(BigSampler.fieldSep.toString))
-        case "BOOLEAN" | "INTEGER" | "FLOAT" | "STRING" | "BYTES" => true
-        case  t => throw new UnsupportedOperationException(
-          s"Type `$t` of field `${field.getName}` is not supported as sampling key")
-      }
-    }
-  }
-
-  private[samplers] def buildKey(schema: => Seq[TableFieldSchema],
-                                 distributionFields: Seq[String])(tr: TableRow)
-  : Set[String] = {
-    distributionFields.map{ f =>
-      getTableRowField(tr, f, schema).toString
-    }.toSet
-  }
-
-  //scalastyle:off method.length cyclomatic.complexity parameter.number
-  // TODO: investigate if possible to move this logic to BQ itself
-  private[samplers] def sample(sc: ScioContext,
-             inputTbl: TableReference,
-             outputTbl: TableReference,
-             fields: List[String],
-             fraction: Double,
-             seed: Option[Int],
-             distribution: Option[SampleDistribution],
-             distributionFields: List[String],
-             precision: Precision,
-             sizePerKey: Int)
-  : Future[Tap[TableRow]] = {
-    import BigQueryIO.Write.CreateDisposition.CREATE_IF_NEEDED
-    import BigQueryIO.Write.WriteDisposition.WRITE_EMPTY
-
-    val patchedBigQueryService = new PatchedBigQueryServicesImpl()
-      .getDatasetService(sc.optionsAs[BigQueryOptions])
-    if (patchedBigQueryService.getTable(outputTbl) != null) {
-      log.info(s"Reuse previous sample at $outputTbl")
-      Taps().bigQueryTable(outputTbl)
-    } else {
-      log.info(s"Will sample from BigQuery table: $inputTbl, output will be $outputTbl")
-      val schema = patchedBigQueryService.getTable(inputTbl).getSchema
-
-      val coll = sc.bigQueryTable(inputTbl)
-
-      val sampledCollection = sampleTableRow(coll, fraction, schema, fields, seed, distribution,
-        distributionFields, precision, sizePerKey)
-
-      val r = sampledCollection
-        .saveAsBigQuery(outputTbl, schema, WRITE_EMPTY, CREATE_IF_NEEDED, tableDescription = "")
-      sc.close().waitUntilDone()
-      r
-    }
-  }
-  //scalastyle:on method.length cyclomatic.complexity parameter.number
-}
